[BE] fix : live starttime 타입 수정

See merge request s06-blockchain-nft-sub2/S06P22E206!200

+++ backend/jpmp/src/main/java/com/jpmp/api/dto/request/live/LiveDto.java
@@ -47,5 +47,5 @@ public class LiveDto {
     private String nftId;
 
     @ApiModelProperty(name = "시작시간", example = "123456")
-    private int starttime;
+    private long starttime;
 }
